AISG/requirements.txt
pandas==0.25.2
numpy==1.18.1
sklearn==0.0
lightgbm==2.3.1
PyYAML==5.1.2